requirements-test.txt
pytest
pytest-asyncio
httpx
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
python-dotenv==1.0.1
asyncpg>=0.29.0
cloud-sql-python-connector[asyncpg]>=1.4.0
SQLAlchemy[asyncio]>=2.0
google-cloud-storage>=2.16.0
google-auth>=2.29.0
motor>=3.4.0